requirements.txt
streamlit==1.330
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchain-classic

openai==1.30.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
pypdf==4.2.0
tiktoken==0.7.0
